:bug: Variable was too small

+++ esp8266.ino
@@ -47,9 +47,9 @@ void handleRoot()
   Serial.println("[WebServer] Request: /");
 
   // Read illuminance (lx)
-  float visible = tsl2561.getLuminosity(TSL2561_VISIBLE);
-  float full = tsl2561.getLuminosity(TSL2561_FULLSPECTRUM);
-  float ir = tsl2561.getLuminosity(TSL2561_INFRARED);
+  uint32_t visible = tsl2561.getLuminosity(TSL2561_VISIBLE);
+  uint32_t full = tsl2561.getLuminosity(TSL2561_FULLSPECTRUM);
+  uint32_t ir = tsl2561.getLuminosity(TSL2561_INFRARED);
 
   // Read temperature (°C)
   float temperature = readTemperature();
